Make _UID and _GID mandatory
Use the Makefile to create a .env file, and alert user if the environment variables are not set. This allows to use docker-compose without the need of Makefile (by creating .env by hand), and to ensure that the var is well created to not change owner:group to the local volumes.